Preserve CLI Hermes codecs across module split
