+++ packages/agent-cli/src/Agent/CLI/Lsp.hs
@@ -1,5 +1,4 @@
--- | Session-owned stdio Language Server Protocol clients for Grok's @lsp@
--- tool. Servers are advertised only after a successful initialize handshake.
+-- | Session-owned stdio Language Server Protocol clients for Grok's @lsp@ tool.
 module Agent.CLI.Lsp
     ( LspRuntime
     , LspStartup(..)
@@ -8,14 +7,27 @@ module Agent.CLI.Lsp
     , lspRuntimeTool
     , encodeLspFrame
     ) where
-
 import Agent.CLI.Config
     ( LspConfig(..)
     , LspServerConfig(..)
     )
+import Agent.CLI.Lsp.Capabilities (clientCapabilities)
 import Agent.CLI.FileUri
     ( fileUri
-    , fileUriPath
+    )
+import Agent.CLI.Lsp.Formatting (formatLspResult)
+import Agent.CLI.Lsp.Path
+    ( exceptionText
+    , pathWithin
+    , quote
+    , resolveWorkspace
+    , sanitizeName
+    )
+import Agent.CLI.Lsp.Protocol
+    ( IncomingMessage(..)
+    , encodeLspFrame
+    , readMessage
+    , sendMessage
     )
 import Agent.Json
     ( RawJson
@@ -24,8 +36,7 @@ import Agent.Json
     , rawJsonFromEncoding
     )
 import Agent.Json.Decode
-    ( JsonError(..)
-    , decodeEither
+    ( decodeEither
     , optionalKey
     )
 import Agent.Json.Decode qualified as Hermes
@@ -52,9 +63,7 @@ import Control.Concurrent.MVar
     , withMVar
     )
 import Control.Exception.Safe
-    ( SomeException
-    , displayException
-    , finally
+    ( finally
     , mask
     , onException
     , tryAny
@@ -68,9 +77,6 @@ import Control.Monad
 import qualified Data.Aeson as Aeson
 import Data.Aeson ((.=))
 import qualified Data.ByteString as BS
-import qualified Data.ByteString.Char8 as BS8
-import qualified Data.ByteString.Lazy as LBS
-import Data.Char (isAlphaNum, toLower)
 import Data.IORef
     ( IORef
     , atomicModifyIORef'
@@ -81,7 +87,7 @@ import Data.IORef
 import Data.List (find)
 import qualified Data.Map.Strict as Map
 import Data.Map.Strict (Map)
-import Data.Maybe (fromMaybe, maybeToList)
+import Data.Maybe (fromMaybe)
 import Data.Text (Text)
 import qualified Data.Text as Text
 import qualified Data.Text.Encoding as Text
@@ -89,7 +95,6 @@ import Data.Text.Encoding.Error (lenientDecode)
 import System.Directory
     ( canonicalizePath
     , createDirectoryIfMissing
-    , doesDirectoryExist
     )
 import System.Environment (getEnvironment)
 import System.Exit (ExitCode)
@@ -117,17 +122,14 @@ import System.Process
     )
 import System.Timeout (timeout)
 import System.Posix.Types (ProcessID)
-
 data LspStartup = LspStartup
     { lspStartupRuntime :: !(Maybe LspRuntime)
     , lspStartupWarnings :: ![Text]
     }
-
 data LspRuntime = LspRuntime
     { runtimeClients :: !(Map Text LspClient)
     , runtimeWorkspace :: !FilePath
     }
-
 data LspClient = LspClient
     { clientName :: !Text
     , clientConfig :: !LspServerConfig
@@ -144,7 +146,6 @@ data LspClient = LspClient
     , clientRequestLock :: !(MVar ())
     , clientClosed :: !(IORef Bool)
     }
-
 newLspRuntime :: LspConfig -> ToolEnv -> IO LspStartup
 newLspRuntime config env
     | not config.lspEnabled =
@@ -234,15 +235,12 @@ newLspRuntime config env
                                         ]
                                     else warnings
                             }
-
 closeLspRuntime :: LspRuntime -> IO ()
 closeLspRuntime runtime =
     mapM_ closeLspClient (Map.elems runtime.runtimeClients)
-
 lspRuntimeTool :: LspRuntime -> AppTool
 lspRuntimeTool runtime =
     lspTool (runLsp runtime)
-
 startLspClient
     :: FilePath
     -> FilePath
@@ -294,7 +292,6 @@ startLspClient workspace logDirectory name config = do
                                     forceCloseLspClient client
                                     pure (name, Left err)
                                 Right () -> pure (name, Right client)
-
 spawnClient
     :: Text
     -> LspServerConfig
@@ -389,7 +386,6 @@ spawnClient name config workspace logHandle = mask \restore -> do
             finishSetup
                 `onException` stopAsync stderrWorker)
         `onException` closePartial
-
 initializeClient :: LspClient -> IO (Either Text ())
 initializeClient client = do
     let rootUri = fileUri client.clientWorkspace
@@ -449,30 +445,6 @@ initializeClient client = do
                                                         ("settings notification failed: "
                                                             <> err))
                                             Right () -> pure (Right ())
-
-clientCapabilities :: Aeson.Value
-clientCapabilities =
-    Aeson.object
-        [ "workspace" .= Aeson.object
-            [ "symbol" .= Aeson.object []
-            , "workspaceFolders" .= True
-            , "configuration" .= True
-            , "applyEdit" .= False
-            ]
-        , "textDocument" .= Aeson.object
-            [ "definition" .= Aeson.object []
-            , "references" .= Aeson.object []
-            , "hover" .= Aeson.object
-                [ "contentFormat" .=
-                    [ "markdown" :: Text
-                    , "plaintext"
-                    ]
-                ]
-            , "implementation" .= Aeson.object []
-            , "documentSymbol" .= Aeson.object []
-            ]
-        ]
-
 runLsp :: LspRuntime -> LspRequest -> IO (Either Text Text)
 runLsp runtime request =
     case request.lspOperation of
@@ -483,7 +455,6 @@ runLsp runtime request =
                 Left err -> pure (Left err)
                 Right (client, path, uri) ->
                     dispatchFileOperation client path uri request operation
-
 runWorkspaceSymbol
     :: LspRuntime
     -> LspRequest
@@ -521,7 +492,6 @@ runWorkspaceSymbol runtime request =
                             <> formatLspResult WorkspaceSymbol value
                         | (name, value) <- successes
                         ]
-
 prepareFileRequest
     :: LspRuntime
     -> LspRequest
@@ -585,7 +555,6 @@ prepareFileRequest runtime request =
                                                 , canonical
                                                 , fileUri canonical
                                                 ))
-
 dispatchFileOperation
     :: LspClient
     -> FilePath
@@ -645,7 +614,6 @@ dispatchFileOperation client path uri request operation =
                 Left err -> pure (Left err)
                 Right value ->
                     pure (Right (formatLspResult operation value))
-
 synchronizeDocument
     :: LspClient
     -> FilePath
@@ -737,28 +705,6 @@ clientForPath clients path =
             && Map.member extension
                 client.clientConfig.lspExtensionToLanguage
 
-data IncomingMessage = IncomingMessage
-    { incomingId :: !(Maybe RawJson)
-    , incomingNumericId :: !(Maybe Int)
-    , incomingMethod :: !(Maybe Text)
-    , incomingParams :: !(Maybe RawJson)
-    , incomingError :: !(Maybe RawJson)
-    , incomingResult :: !(Maybe RawJson)
-    }
-
-incomingMessageDecoder :: Hermes.Decoder IncomingMessage
-incomingMessageDecoder =
-    Hermes.object do
-        identifier <- optionalKey "id" rawJsonDecoder
-        IncomingMessage
-            identifier
-            (identifier >>= either (const Nothing) Just
-                . decodeRawJson Hermes.int)
-            <$> optionalKey "method" Hermes.text
-            <*> optionalKey "params" rawJsonDecoder
-            <*> optionalKey "error" rawJsonDecoder
-            <*> optionalKey "result" rawJsonDecoder
-
 configurationCountDecoder :: Hermes.Decoder Int
 configurationCountDecoder =
     Hermes.object $
@@ -767,7 +713,7 @@ configurationCountDecoder =
 
 decodeRawJson :: Hermes.Decoder a -> RawJson -> Either Text a
 decodeRawJson decoder =
-    either (Left . jsonErrorMessage) Right
+    either (Left . Hermes.jsonErrorMessage) Right
         . decodeEither decoder
         . rawJsonBytes
 
@@ -855,6 +801,10 @@ awaitResponse client requestId =
                 awaitResponse client requestId
             | otherwise ->
                 awaitResponse client requestId
+
+compactJson :: RawJson -> Text
+compactJson =
+    Text.decodeUtf8With lenientDecode . rawJsonBytes
 
 answerServerRequest
     :: LspClient
@@ -954,106 +904,6 @@ sendMessageWithin client timeoutMilliseconds label value =
                         <> exceptionText exception
             Just (Right ()) -> pure (Right ())
 
-sendMessage :: Handle -> Aeson.Value -> IO ()
-sendMessage handle value = do
-    BS.hPut handle (encodeLspFrame value)
-    hFlush handle
-
-encodeLspFrame :: Aeson.Value -> BS.ByteString
-encodeLspFrame value =
-    let body = LBS.toStrict (Aeson.encode value)
-    in BS8.pack
-        ("Content-Length: " <> show (BS.length body) <> "\r\n\r\n")
-            <> body
-
-readMessage :: Handle -> IO (Either Text IncomingMessage)
-readMessage handle = do
-    headers <- tryAny (readHeaders handle 0 0 Map.empty)
-    case headers of
-        Left exception ->
-            pure . Left $
-                "failed to read LSP response headers: "
-                    <> exceptionText exception
-        Right values ->
-            case Map.lookup "content-length" values >>= readMaybeInt of
-                Nothing ->
-                    pure (Left "LSP response omitted a valid Content-Length")
-                Just bodyLength
-                    | bodyLength < 0 ->
-                        pure (Left "LSP response had a negative Content-Length")
-                    | bodyLength > maxLspMessageBytes ->
-                        pure . Left $
-                            "LSP response exceeds "
-                                <> Text.pack (show maxLspMessageBytes)
-                                <> " bytes"
-                    | otherwise -> do
-                        bodyResult <- tryAny (BS.hGet handle bodyLength)
-                        pure case bodyResult of
-                            Left exception ->
-                                Left
-                                    ( "failed to read LSP response body: "
-                                        <> exceptionText exception
-                                    )
-                            Right body
-                                | BS.length body /= bodyLength ->
-                                    Left "LSP response ended before Content-Length"
-                                | otherwise ->
-                                    case decodeEither incomingMessageDecoder body of
-                                        Left (JsonError err) ->
-                                            Left
-                                                ( "LSP response was invalid JSON: "
-                                                    <> err
-                                                )
-                                        Right value -> Right value
-
-readHeaders
-    :: Handle
-    -> Int
-    -> Int
-    -> Map String String
-    -> IO (Map String String)
-readHeaders handle count totalBytes headers = do
-    when (count >= maxLspHeaderCount) $
-        ioError (userError "LSP response sent too many headers")
-    when (totalBytes >= maxLspHeaderBytes) $
-        ioError (userError "LSP response headers are too large")
-    rawLine <- BS8.hGetLine handle
-    let line = BS8.unpack (BS8.takeWhile (/= '\r') rawLine)
-        nextBytes = totalBytes + BS.length rawLine
-    when (nextBytes > maxLspHeaderBytes) $
-        ioError (userError "LSP response headers are too large")
-    if null line
-        then pure headers
-        else
-            case break (== ':') line of
-                (name, ':' : value) ->
-                    readHeaders handle (count + 1) nextBytes $
-                        Map.insert
-                            (map toLower name)
-                            (dropWhile (== ' ') value)
-                            headers
-                _ ->
-                    readHeaders
-                        handle
-                        (count + 1)
-                        nextBytes
-                        headers
-
-maxLspHeaderCount :: Int
-maxLspHeaderCount = 100
-
-maxLspHeaderBytes :: Int
-maxLspHeaderBytes = 64 * 1024
-
-maxLspMessageBytes :: Int
-maxLspMessageBytes = 16 * 1024 * 1024
-
-readMaybeInt :: String -> Maybe Int
-readMaybeInt raw =
-    case reads raw of
-        [(value, "")] -> Just value
-        _ -> Nothing
-
 closeLspClient :: LspClient -> IO ()
 closeLspClient client =
     withMVar client.clientRequestLock \() -> do
@@ -1150,213 +1000,3 @@ stopAsync worker =
 ignoreException :: IO a -> IO ()
 ignoreException action =
     void (tryAny action)
-
-resolveWorkspace
-    :: FilePath
-    -> Maybe Text
-    -> IO (Either Text FilePath)
-resolveWorkspace workspace override = do
-    let requested = case override of
-            Nothing -> workspace
-            Just value
-                | FilePath.isAbsolute (Text.unpack value) ->
-                    Text.unpack value
-                | otherwise ->
-                    workspace FilePath.</> Text.unpack value
-    resolved <-
-        tryAny do
-            exists <- doesDirectoryExist requested
-            unless exists $
-                ioError (userError "workspace folder does not exist")
-            canonicalizePath requested
-    canonicalWorkspace <- tryAny (canonicalizePath workspace)
-    pure case (canonicalWorkspace, resolved) of
-        (Left exception, _) ->
-            Left
-                ( "failed to resolve active workspace: "
-                    <> exceptionText exception
-                )
-        (_, Left exception) ->
-            Left
-                ( "invalid workspaceFolder: "
-                    <> exceptionText exception
-                )
-        (Right root, Right child)
-            | pathWithin root child -> Right child
-            | otherwise ->
-                Left "workspaceFolder must be inside the active workspace"
-
-pathWithin :: FilePath -> FilePath -> Bool
-pathWithin root candidate =
-    let relative =
-            FilePath.normalise
-                (FilePath.makeRelative
-                    (FilePath.normalise root)
-                    (FilePath.normalise candidate))
-    in relative == "."
-        || (not (FilePath.isAbsolute relative)
-            && relative /= ".."
-            && not
-                ( (".." <> [FilePath.pathSeparator])
-                    `isPrefixOfString` relative
-                ))
-
-formatLspResult :: LspOperation -> RawJson -> Text
-formatLspResult operation value =
-    case operation of
-        GoToDefinition -> formatLocations "definition" value
-        FindReferences -> formatLocations "references" value
-        GoToImplementation -> formatLocations "implementations" value
-        Hover -> formatHover value
-        DocumentSymbol -> formatSymbols value
-        WorkspaceSymbol -> formatSymbols value
-
-formatLocations :: Text -> RawJson -> Text
-formatLocations label value =
-    case decodeRawJson locationsDecoder value of
-        Left _ -> compactJson value
-        Right [] -> "No " <> label <> " found."
-        Right locations -> Text.intercalate "\n" locations
-
-locationsDecoder :: Hermes.Decoder [Text]
-locationsDecoder =
-    Hermes.getType >>= \case
-        Hermes.VArray ->
-            concat <$> Hermes.list locationsDecoder
-        Hermes.VObject ->
-            maybeToList <$> locationDecoder
-        _ -> [] <$ rawJsonDecoder
-
-locationDecoder :: Hermes.Decoder (Maybe Text)
-locationDecoder =
-    Hermes.object do
-        uriValue <- optionalKey "uri" Hermes.text
-        targetUri <- optionalKey "targetUri" Hermes.text
-        rangeValue <- optionalKey "range" startPositionDecoder
-        selectionRange <-
-            optionalKey "targetSelectionRange" startPositionDecoder
-        targetRange <- optionalKey "targetRange" startPositionDecoder
-        let uri = uriValue <|> targetUri
-            range = rangeValue <|> selectionRange <|> targetRange
-        pure do
-            locationUri <- uri
-            let (line, character) = fromMaybe (0, 0) range
-                path = maybe locationUri Text.pack (fileUriPath locationUri)
-            pure $
-                path
-                    <> ":"
-                    <> Text.pack (show (line + 1))
-                    <> ":"
-                    <> Text.pack (show (character + 1))
-
-startPositionDecoder :: Hermes.Decoder (Int, Int)
-startPositionDecoder =
-    Hermes.object $
-        Hermes.atKey "start" $
-            Hermes.object $
-                (,)
-                    <$> Hermes.atKey "line" Hermes.int
-                    <*> Hermes.atKey "character" Hermes.int
-
-formatHover :: RawJson -> Text
-formatHover value =
-    either (const (compactJson value)) id $
-        decodeRawJson hoverDecoder value
-
-hoverDecoder :: Hermes.Decoder Text
-hoverDecoder =
-    Hermes.getType >>= \case
-        Hermes.VNull ->
-            "No hover information found." <$ Hermes.isNull
-        Hermes.VString -> Hermes.text
-        Hermes.VArray ->
-            Text.intercalate "\n\n" <$> Hermes.list hoverDecoder
-        Hermes.VObject ->
-            Hermes.withRawJsonByteString \bytes ->
-                pure $
-                    either
-                        (const (Text.decodeUtf8With lenientDecode bytes))
-                        (fromMaybe (Text.decodeUtf8With lenientDecode bytes))
-                        (decodeEither hoverObjectDecoder (BS.copy bytes))
-        _ ->
-            rawTextDecoder
-
-hoverObjectDecoder :: Hermes.Decoder (Maybe Text)
-hoverObjectDecoder =
-    Hermes.object do
-        contents <- optionalKey "contents" hoverDecoder
-        value <- optionalKey "value" Hermes.text
-        language <- optionalKey "language" Hermes.text
-        pure (contents <|> value <|> language)
-
-formatSymbols :: RawJson -> Text
-formatSymbols value =
-    case decodeRawJson (symbolLinesDecoder 0) value of
-        Left _ -> compactJson value
-        Right [] -> "No symbols found."
-        Right lines' -> Text.intercalate "\n" lines'
-
-symbolLinesDecoder :: Int -> Hermes.Decoder [Text]
-symbolLinesDecoder depth =
-    Hermes.getType >>= \case
-        Hermes.VArray ->
-            concat <$> Hermes.list (symbolLinesDecoder depth)
-        Hermes.VObject ->
-            Hermes.object do
-                name <- optionalKey "name" Hermes.text
-                location <-
-                    optionalKey "location" locationDecoder
-                directLocation <- Hermes.liftObjectDecoder locationDecoder
-                children <-
-                    fromMaybe []
-                        <$> optionalKey "children"
-                            (symbolLinesDecoder (depth + 1))
-                pure case name of
-                    Nothing -> []
-                    Just symbolName ->
-                        let suffix =
-                                maybe ""
-                                    (" — " <>)
-                                    ((location >>= id) <|> directLocation)
-                            current =
-                                Text.replicate depth "  "
-                                    <> "- "
-                                    <> symbolName
-                                    <> suffix
-                        in current : children
-        _ -> [] <$ rawJsonDecoder
-
-rawTextDecoder :: Hermes.Decoder Text
-rawTextDecoder =
-    Hermes.withRawJsonByteString $
-        pure . Text.decodeUtf8With lenientDecode
-
-compactJson :: RawJson -> Text
-compactJson =
-    Text.decodeUtf8With lenientDecode . rawJsonBytes
-
-sanitizeName :: Text -> FilePath
-sanitizeName =
-    Text.unpack
-        . Text.map
-            (\character ->
-                if isAlphaNum character || character `elem` ("-_" :: String)
-                    then character
-                    else '_')
-
-exceptionText :: SomeException -> Text
-exceptionText = Text.pack . displayException
-
-quote :: Text -> Text
-quote value = "'" <> value <> "'"
-
-isPrefixOfString :: String -> String -> Bool
-isPrefixOfString prefix value =
-    take (length prefix) value == prefix
-
-infixr 3 <|>
-
-(<|>) :: Maybe a -> Maybe a -> Maybe a
-first <|> second = case first of
-    Just value -> Just value
-    Nothing -> second
